added loop.py and me.py

+++ area.py
@@ -1,9 +1,9 @@
 radius=input(print("Enter the radius of the circle: "))
 height=input(print("Enter the height of the cyclinder: "))
 pi=3.14
-diameter=int(radius)*2
-area=(2*pi*int(radius)*int(radius))
-area2=(pi*diameter*int(height))
+diameter=float(radius)*2
+area=(2*pi*float(radius)*float(radius))
+area2=(pi*diameter*float(height))
 print(area+area2)
